- TSYS payment request added

+++ app/src/main/java/com/android/pos/ui/fragments/checkout/CheckoutDineInFragmentNew.kt
@@ -1172,7 +1172,7 @@ class CheckoutDineInFragmentNew(val dineInDataModel: CheckOutDineInDataModel) : 
     ) {
 
         val jsonArray1 = magtekRequestUtils.processManualEntry(
-            (paymentAmount * 100).toInt(),
+            (paymentAmount * 100),
             cardNumber,
             expDate,
             cardCVV
@@ -1869,7 +1869,7 @@ class CheckoutDineInFragmentNew(val dineInDataModel: CheckOutDineInDataModel) : 
 
         val jsonArray1 = magtekModule.m_scra?.let {
             magtekRequestUtils.processCardSwipe(
-                (paymentAmount * 100).toInt(),
+                (paymentAmount * 100),
                 magtekModule.m_scra!!.ksn,
                 magtekModule.m_scra!!.magnePrint,
                 magtekModule.m_scra!!.magnePrintStatus,
@@ -1973,7 +1973,7 @@ class CheckoutDineInFragmentNew(val dineInDataModel: CheckOutDineInDataModel) : 
         ProgressUtils.dismissProgressDialog()
 
         val jsonArray1 = magtekRequestUtils.processData(
-            (paymentAmount * 100).toInt(),
+            (paymentAmount * 100),
             TLVParser.getHexString(data),
             Constants.AUTHORIZE
         )
@@ -2030,7 +2030,7 @@ class CheckoutDineInFragmentNew(val dineInDataModel: CheckOutDineInDataModel) : 
                     dismissDialog()
 
                     val jsonArray1 = magtekRequestUtils.processData(
-                        (paymentAmount * 100).toInt(),
+                        (paymentAmount * 100),
                         MTParser.getHexString(data.ByteArray()),
                         Constants.AUTHORIZE
                     )
